requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
pyaes==1.6.1
Pyrogram==2.0.24
PySocks==1.7.1
python-dotenv==0.20.0
requests==2.27.1
TgCrypto==1.2.3
urllib3==1.26.9
